Add files via upload
The simulation configurations should be specified in an input file sim_config.txt that the program reads. The program should not prompt the user for any input. The input file should contain the following values in the exact order specified, one per line:
- Duration of simulation in seconds (valid range is 3600-36000).
- Number of simulation experiments (valid range is 1-100. In the existing program, this value
is set to 10).
- Minimum task size (must be in range 1-100)
- Maximum task size (must be in range 1-100 and >= minimum)
- Number of printers (must be 1 or 2)
- Page rate of printer 1 (must be in the range 1-50)
- Page rate of printer 2 (if number of printers is 1, this value would not be specified)
The input file contains 6 or 7 integers, depending on whether 1 or 2 printers were specified.
Your program must implement input validation (Using exception handling and conditional statements as appropriate) to detect format error and out of range errors. Appropriate error messages should be output to the terminal, after which point the program should terminate.